refactor: next/script 사용하는 방식으로 변경
